deleted classes.py, added a classes/ folder that contains state.py config.py weather.py and time.py, rerouted all references to those classes
