+++ requirements.txt
@@ -13,3 +13,4 @@ typing_extensions
 sphinx
 sphinx-rtd-theme
 juliandate
+salib
